# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoiça API",
        "description": "API de gestion de devis et factures pour artisans.",
        "contact": {
            "email": "contact@invoica.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Développement"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/v1/company": {
            "get": {
                "tags": [
                    "Entreprise"
                ],
                "summary": "Afficher l'entreprise",
                "operationId": "0c0749190933d758b3c738be590788a8",
                "responses": {
                    "200": {
                        "description": "Détails de l'entreprise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune entreprise trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Entreprise"
                ],
                "summary": "Modifier l'entreprise",
                "operationId": "6d07902996eba7aad243758a82de1fba",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "manager_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entreprise mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Entreprise"
                ],
                "summary": "Créer l'entreprise",
                "operationId": "d5834f135441d2027f02cd07d64e8b38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Brico Pro"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "10 rue de Paris"
                                    },
                                    "manager_name": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "signature": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "stamp": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entreprise créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Entreprise déjà existante"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Company": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string"
                    },
                    "manager_name": {
                        "type": "string"
                    },
                    "signature": {
                        "type": "string",
                        "nullable": true
                    },
                    "stamp": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Entrez votre token Sanctum (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}